Update test_hard_negatives test, simplify mine_hard_negatives slightly
